[bugfix/feature] Fix control handling in subsetMIDAS.
Add an argument filter_controls to subsetMIDAS to also filter control
rows. This is important when filtering cell lines as the information
provided for the subsetting is not used when using MIDAS, only ID.type
is.
However the controls should still not be filtered when looking for a
subset of treatments.